// web/src/api/model/bookResponseDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Rust Web App
 * OpenAPI spec version: 0.1.0
 */
import type { BookResponseDtoAppliedAt } from './bookResponseDtoAppliedAt';
import type { BookPublisherDto } from './bookPublisherDto';
import type { BookResponseDtoShop } from './bookResponseDtoShop';

export interface BookResponseDto {
  applied_at?: BookResponseDtoAppliedAt;
  author: string;
  format: string;
  price: number;
  pub_id: string;
  publisher: BookPublisherDto;
  shop?: BookResponseDtoShop;
  title: string;
}