// web/src/api/model/bookUpdateDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Rust Web App
 * OpenAPI spec version: 0.1.0
 */
import type { BookUpdateDtoFormat } from './bookUpdateDtoFormat';
import type { BookUpdateDtoShopId } from './bookUpdateDtoShopId';

export interface BookUpdateDto {
  author: string;
  format?: BookUpdateDtoFormat;
  price: number;
  publisher_id: string;
  shop_id?: BookUpdateDtoShopId;
  title: string;
}